system: tighten provider effect boundaries ( spec001w6 )
